A number of changes in order to address the deadlocking issues raised by Issue #2:
- Made "incremental listeners" array in GtfsRealtimeExporterImpl
threadsafe via CopyOnWriteArrayList.  This allows us to remove the
sychronization around add/removing listeners.  The assumption here is
that most users of the library will be handling a reasonably small
number of clients that don't change that often.  If that assumption
doesn't hold, we can revisit.
- Added some synchronization + cleanup in the actual data socket, in
order to more gracefully handle concurrent updates and disconnects.
- Added an integration test that attempts to stress the system with a
high volume of incremental updates and frequent disconnect/connect by
the websocket client.